third_party/WebKit/LayoutTests/external/wpt/html/semantics/scripting-1/the-script-element/execution-timing/128-expected.txt

This is a testharness.js-based test.
FAIL scheduler: appending script element to script  assert_array_equals: property 1, expected "inline script #3" but got "inline script #2"
Harness: the test ran to completion.
